fix(ViewProxy): expose initial camera properties in ViewProxy
Migrate `rotate` and `updateOrientation` from View2DProxy to ViewProxy.
