+++ Sources/Proxy/Core/View2DProxy/index.js
@@ -1,5 +1,4 @@
 import macro from 'vtk.js/Sources/macro';
-import vtkMatrixBuilder from 'vtk.js/Sources/Common/Core/MatrixBuilder';
 
 import vtkViewProxy from 'vtk.js/Sources/Proxy/Core/ViewProxy';
 
@@ -11,14 +10,9 @@ function vtkView2DProxy(publicAPI, model) {
   // Set our className
   model.classHierarchy.push('vtkView2DProxy');
 
+  const superUpdateOrientation = publicAPI.updateOrientation;
   publicAPI.updateOrientation = (axisIndex, orientation, viewUp) => {
-    model.axis = axisIndex;
-    model.orientation = orientation;
-    model.viewUp = viewUp;
-    const position = model.camera.getFocalPoint();
-    position[model.axis] += model.orientation;
-    model.camera.setPosition(...position);
-    model.camera.setViewUp(...viewUp);
+    superUpdateOrientation(axisIndex, orientation, viewUp);
 
     let count = model.representations.length;
     while (count--) {
@@ -31,30 +25,6 @@ function vtkView2DProxy(publicAPI, model) {
 
     publicAPI.updateCornerAnnotation({ axis: 'XYZ'[axisIndex] });
   };
-
-  publicAPI.rotate = (angle) => {
-    const { viewUp, focalPoint, position } = model.camera.get(
-      'viewUp',
-      'focalPoint',
-      'position'
-    );
-    const axis = [
-      focalPoint[0] - position[0],
-      focalPoint[1] - position[1],
-      focalPoint[2] - position[2],
-    ];
-
-    vtkMatrixBuilder
-      .buildFromDegree()
-      .rotate(Number.isNaN(angle) ? 90 : angle, axis)
-      .apply(viewUp);
-
-    model.camera.setViewUp(...viewUp);
-    model.camera.modified();
-    model.renderWindow.render();
-  };
-
-  publicAPI.updateOrientation(model.axis, model.orientation, model.viewUp);
 
   // Setup default corner annotation
   /* eslint-disable no-template-curly-in-string */
@@ -79,9 +49,6 @@ function vtkView2DProxy(publicAPI, model) {
 // ----------------------------------------------------------------------------
 
 const DEFAULT_VALUES = {
-  axis: 2,
-  orientation: -1,
-  viewUp: [0, 1, 0],
   useParallelRendering: true,
 };
