+++ Sources/Proxy/Core/ViewProxy/index.js
@@ -3,6 +3,7 @@ import macro from 'vtk.js/Sources/macro';
 import vtkAnnotatedCubeActor from 'vtk.js/Sources/Rendering/Core/AnnotatedCubeActor';
 import vtkCornerAnnotation from 'vtk.js/Sources/Interaction/UI/CornerAnnotation';
 import vtkInteractorStyleManipulator from 'vtk.js/Sources/Interaction/Style/InteractorStyleManipulator';
+import vtkMatrixBuilder from 'vtk.js/Sources/Common/Core/MatrixBuilder';
 import vtkOpenGLRenderWindow from 'vtk.js/Sources/Rendering/OpenGL/RenderWindow';
 import vtkOrientationMarkerWidget from 'vtk.js/Sources/Interaction/Widgets/OrientationMarkerWidget';
 import vtkRenderer from 'vtk.js/Sources/Rendering/Core/Renderer';
@@ -267,6 +268,43 @@ function vtkViewProxy(publicAPI, model) {
       model.renderWindow.getInteractor().cancelAnimation('proxy');
     }
   };
+
+  // --------------------------------------------------------------------------
+
+  publicAPI.updateOrientation = (axisIndex, orientation, viewUp) => {
+    model.axis = axisIndex;
+    model.orientation = orientation;
+    model.viewUp = viewUp;
+    const position = model.camera.getFocalPoint();
+    position[model.axis] += model.orientation;
+    model.camera.setPosition(...position);
+    model.camera.setViewUp(...viewUp);
+  };
+  publicAPI.updateOrientation(model.axis, model.orientation, model.viewUp);
+
+  // --------------------------------------------------------------------------
+
+  publicAPI.rotate = (angle) => {
+    const { viewUp, focalPoint, position } = model.camera.get(
+      'viewUp',
+      'focalPoint',
+      'position'
+    );
+    const axis = [
+      focalPoint[0] - position[0],
+      focalPoint[1] - position[1],
+      focalPoint[2] - position[2],
+    ];
+
+    vtkMatrixBuilder
+      .buildFromDegree()
+      .rotate(Number.isNaN(angle) ? 90 : angle, axis)
+      .apply(viewUp);
+
+    model.camera.setViewUp(...viewUp);
+    model.camera.modified();
+    model.renderWindow.render();
+  };
 }
 
 // ----------------------------------------------------------------------------
@@ -278,6 +316,9 @@ const DEFAULT_VALUES = {
   sectionName: 'view',
   annotationOpacity: 1,
   resetCameraOnFirstRender: true,
+  axis: 2,
+  orientation: -1,
+  viewUp: [0, 1, 0],
 };
 
 // ----------------------------------------------------------------------------
